Project 2 complete.
NOTE: contains folder menu that has 2 functions I use for making a gui menu.

Write a menu driven application that is going to use the hierarchy that you have written for homework 8 and the functions that you have written for homework 9. You will keep track of the collection of Files in a vector (File pointers, since File is an abstract class). The menu should contain the following options:
    1. Read from a file (Read descriptions of Files from "files.txt" creating appropriate File object and putting it inside your vector. The file.txt is provided)
    2. Create an Image File (Creates a single Image File asking the user for its properties)
    3. Create a Text File (Creates a single Text File asking the user for its properties)
    4. Print All Files (You must use functions written for homework 9)
    5. Print Image Files (You must use functions written for homework 9)
    6. Print Text Files (You must use functions written for homework 9)
    7. Delete a file by file name and type (Removes a file from a vector)
    8. Quit (Terminate the program)